server/requirements.txt
fastapi
uvicorn
pydantic
psycopg2-binary
sqlalchemy
paho-mqtt
passlib[bcrypt]
python-dotenv
pandas